% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF2MAF.R
\name{convert_vcfs_to_maf}
\alias{convert_vcfs_to_maf}
\title{Convert several VCF files into a cohort-level MAF}
\usage{
convert_vcfs_to_maf(
  vcfs,
  ref_genome,
  tumor_id = paste0("Tumor", seq_len(length(vcfs)), times = length(vcfs)),
  vcf_tumor_id = "TUMOR",
  vcf_normal_id = "NORMAL",
  parse_tumor_id_from_filename = TRUE,
  extract = c("before_dot", "before_underscore"),
  verbose = TRUE
)
}
\arguments{
\item{vcfs}{path to vcf files (character)}

\item{ref_genome}{name of the reference genome used to call variants (string)}

\item{tumor_id}{desired value of Tumor_Sample_Barcode in maf. By default will use the name of the tumor sample in the VCF (string)}

\item{vcf_tumor_id}{a vector describing what the tumor_names to expect in a VCF are}

\item{vcf_normal_id}{the sample ID describing the normal in the (string)}

\item{parse_tumor_id_from_filename}{should tumour id be parsed from filename (flag)}

\item{extract}{what text do we extract as sample name. See details}

\item{verbose}{(flag)}
}
\value{
a maf-compatible data.frame
}
\description{
Take multiple VCF files, each representing a single tumour-normal file and combine into one big maf
}
\details{
Re-export of \code{\link[vcf2mafR:vcfs2maf]{vcf2mafR::vcfs2maf()}}
}
