% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRUX.R
\name{gistic_files_to_crux}
\alias{gistic_files_to_crux}
\title{Gistic to CRUX}
\usage{
gistic_files_to_crux(
  gisticAllLesionsFile,
  gisticAmpGenesFile,
  gisticDelGenesFile,
  gisticScoresFile,
  outfile,
  cnLevel = c("all", "deep", "shallow"),
  isTCGA = FALSE,
  verbose = TRUE,
  compress = c("gzip", "bzip2", "xz", "none")
)
}
\arguments{
\item{gisticAllLesionsFile}{All Lesions file generated by gistic. e.g; all_lesions.conf_XX.txt, where XX is the confidence level. Required. Default NULL.}

\item{gisticAmpGenesFile}{Amplification Genes file generated by gistic. e.g; amp_genes.conf_XX.txt, where XX is the confidence level. Default NULL.}

\item{gisticDelGenesFile}{Deletion Genes file generated by gistic. e.g; del_genes.conf_XX.txt, where XX is the confidence level. Default NULL.}

\item{gisticScoresFile}{scores.gistic file generated by gistic.}

\item{outfile}{outfile path}

\item{cnLevel}{level of CN changes to use. Can be 'all', 'deep' or 'shallow'. Default uses all i.e, genes with both 'shallow' or 'deep' CN changes}

\item{isTCGA}{Is the data from TCGA. Default FALSE.}

\item{verbose}{Default TRUE}

\item{compress}{compression method}
}
\value{
NULL, invisibly
}
\description{
Create Crux-compatible GISTIC RDS file
}
