% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRUX.R
\name{gistic_folder_to_crux}
\alias{gistic_folder_to_crux}
\title{Gistic Folder to CRUX}
\usage{
gistic_folder_to_crux(gistic_folder, outfile, ...)
}
\arguments{
\item{gistic_folder}{path to a folder containing gistic output files}

\item{outfile}{outfile path}

\item{...}{
  Arguments passed on to \code{\link[=gistic_files_to_crux]{gistic_files_to_crux}}
  \describe{
    \item{\code{compress}}{compression method}
    \item{\code{gisticAllLesionsFile}}{All Lesions file generated by gistic. e.g; all_lesions.conf_XX.txt, where XX is the confidence level. Required. Default NULL.}
    \item{\code{gisticAmpGenesFile}}{Amplification Genes file generated by gistic. e.g; amp_genes.conf_XX.txt, where XX is the confidence level. Default NULL.}
    \item{\code{gisticDelGenesFile}}{Deletion Genes file generated by gistic. e.g; del_genes.conf_XX.txt, where XX is the confidence level. Default NULL.}
    \item{\code{gisticScoresFile}}{scores.gistic file generated by gistic.}
    \item{\code{cnLevel}}{level of CN changes to use. Can be 'all', 'deep' or 'shallow'. Default uses all i.e, genes with both 'shallow' or 'deep' CN changes}
    \item{\code{isTCGA}}{Is the data from TCGA. Default FALSE.}
    \item{\code{verbose}}{Default TRUE}
  }}
}
\value{
NULL, invisibly.
}
\description{
Create Crux-compatible GISTIC RDS file
}
