% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRUX.R
\name{gistic_tar_to_crux}
\alias{gistic_tar_to_crux}
\title{Gistic Zip to CRUX}
\usage{
gistic_tar_to_crux(gistic_tar, outfile, ...)
}
\arguments{
\item{outfile}{path to write CRUX RDS to}

\item{gistic_zip}{path to gistic zip file}

\item{inheritDotParams}{gistic_files_to_crux}
}
\value{
Run for its side effects (creation of an RDS crux gistic at path 'outfile')
}
\description{
Create Crux-compatible GISTIC RDS file
}
