% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interchange-package.R
\docType{package}
\name{interchange-package}
\alias{interchange}
\alias{interchange-package}
\title{interchange: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/selkamand/interchange}
  \item Report bugs at \url{https://github.com/selkamand/interchange/issues}
}

}
\author{
\strong{Maintainer}: Sam El-Kamand \email{selkamand@ccia.org.au} (\href{https://orcid.org/0000-0003-2270-8088}{ORCID})

}
\keyword{internal}
