% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available.R
\name{supported_conversions}
\alias{supported_conversions}
\title{List supported file format conversions}
\usage{
supported_conversions()
}
\value{
A dataframe with two columns: \code{Input} and \code{Output}, describing the source and target file formats for each conversion.
}
\description{
List supported file format conversions
}
\examples{
supported_conversions()
}
